teensy3:  add SECURE_MODE #define to allow for a compiler switch to turn on flash security.
Currently, there is no automated way to turn on flash security from
the Arduino UI.  This is a little bit of a headache because whenever
you want to enable or disable the flash security settings, you must go
and edit teensy3/mk20dx128.c and change the value of FSEC.  Not the
end of the world, but not terribly friendly either.

This change adds compiler define that lets you set the switch from a
maybe from a setting in the boards.txt file.

By adding the following to boards.txt you get a menu item that allows
you to easily switch between secure and insecure modes.

+menu.secure=Secure Flash
+
+teensy36.menu.secure.insecureopt=Insecure
+teensy36.menu.secure.secureopt=Secure
+teensy35.menu.secure.insecureopt=Insecure
+teensy35.menu.secure.secureopt=Secure
+teensy31.menu.secure.insecureopt=Insecure
+teensy31.menu.secure.secureopt=Secure
+teensy36.menu.secure.secureopt.build.flags.c=-DTEENSY3_SECURE_MODE
+teensy35.menu.secure.secureopt.build.flags.c=-DTEENSY3_SECURE_MODE
+teensy31.menu.secure.secureopt.build.flags.c=-DTEENSY3_SECURE_MODE